requirements.txt
playwright==1.44.*
httpx==0.27.*
pyyaml==6.0.*
python-dotenv==1.0.*
notion-client==2.2.*
line-bot-sdk==3.10.*
geopy==2.4.*
jinja2==3.1.*
